Extract shared utils to modules
