+++ src/desktop_release.ts
@@ -16,133 +16,21 @@ limitations under the License.
 
 import { promises as fsProm } from 'fs';
 import * as path from 'path';
-import * as childProcess from 'child_process';
 
 import * as rimraf from 'rimraf';
 
 import getSecret from './get_secret';
 import GitRepo from './gitrepo';
 import logger from './logger';
-
 import Runner, { IRunner } from './runner';
 import DockerRunner from './docker_runner';
-
 import WindowsBuilder from './windows_builder';
 import { ENABLED_TARGETS, Target, TargetId, WindowsTarget } from './target';
+import { setDebVersion, pullDebDatabase, pushDebDatabase, addDeb } from './debian';
+import { getMatchingFilesInDir, pullArtifacts, pushArtifacts, copyAndLog } from './artifacts';
 
 const DESKTOP_GIT_REPO = 'https://github.com/vector-im/element-desktop.git';
 const ELECTRON_BUILDER_CFG_FILE = 'electron-builder.json';
-
-async function setDebVersion(ver, templateFile, outFile) {
-    // Create a debian package control file with the version.
-    // We use a custom control file so we need to do this ourselves
-    let contents = await fsProm.readFile(templateFile, 'utf8');
-    contents += 'Version: ' + ver + "\n";
-    await fsProm.writeFile(outFile, contents);
-
-    logger.info("Version set to " + ver);
-}
-
-async function getMatchingFilesInDir(dir, exp) {
-    const ret = [];
-    for (const f of await fsProm.readdir(dir)) {
-        if (exp.test(f)) {
-            ret.push(f);
-        }
-    }
-    if (ret.length === 0) throw new Error("No files found matching " + exp.toString() + "!");
-    return ret;
-}
-
-async function getRepoTargets(repoDir) {
-    const confDistributions = await fsProm.readFile(path.join(repoDir, 'conf', 'distributions'), 'utf8');
-    const ret = [];
-    for (const line of confDistributions.split('\n')) {
-        if (line.startsWith('Codename')) {
-            ret.push(line.split(': ')[1]);
-        }
-    }
-    return ret;
-}
-
-function pullDebDatabase(debDir, rsyncRoot) {
-    logger.info("Pulling debian database...", rsyncRoot + 'debian/', debDir);
-    return new Promise((resolve, reject) => {
-        const proc = childProcess.spawn('rsync', [
-            '-av', '--delete', rsyncRoot + 'debian/', debDir,
-        ], {
-            stdio: 'inherit',
-        });
-        proc.on('exit', code => {
-            code ? reject(code) : resolve();
-        });
-    });
-}
-
-function pushDebDatabase(debDir, rsyncRoot) {
-    logger.info("Pushing debian database...");
-    return new Promise((resolve, reject) => {
-        const proc = childProcess.spawn('rsync', [
-            '-av', '--delete', debDir + '/', rsyncRoot + 'debian',
-        ], {
-            stdio: 'inherit',
-        });
-        proc.on('exit', code => {
-            code ? reject(code) : resolve();
-        });
-    });
-}
-
-async function addDeb(debDir, deb) {
-    const targets = await getRepoTargets(debDir);
-    logger.info("Adding " + deb + " for " + targets.join(', ') + "...");
-    for (const target of targets) {
-        await new Promise((resolve, reject) => {
-            const proc = childProcess.spawn('reprepro', [
-                'includedeb', target, deb,
-            ], {
-                stdio: 'inherit',
-                cwd: debDir,
-            });
-            proc.on('exit', code => {
-                code ? reject(code) : resolve();
-            });
-        });
-    }
-}
-
-function pullArtifacts(pubDir, rsyncRoot) {
-    logger.info("Pulling artifacts...");
-    return new Promise((resolve, reject) => {
-        const proc = childProcess.spawn('rsync', [
-            '-av', '--delete', rsyncRoot + 'packages.riot.im/', pubDir,
-        ], {
-            stdio: 'inherit',
-        });
-        proc.on('exit', code => {
-            code ? reject(code) : resolve();
-        });
-    });
-}
-
-function pushArtifacts(pubDir, rsyncRoot) {
-    logger.info("Uploading artifacts...");
-    return new Promise((resolve, reject) => {
-        const proc = childProcess.spawn('rsync', [
-            '-av', '--delete', '--delay-updates', pubDir + '/', rsyncRoot + 'packages.riot.im',
-        ], {
-            stdio: 'inherit',
-        });
-        proc.on('exit', code => {
-            code ? reject(code) : resolve();
-        });
-    });
-}
-
-function copyAndLog(src, dest) {
-    logger.info('Copy ' + src + ' -> ' + dest);
-    return fsProm.copyFile(src, dest);
-}
 
 class DesktopReleaseBuilder {
     constructor(winVmName, winUsername, winPassword, rsyncRoot, desktopBranch) {
